Factor DMRoomTile out to its own file
I'm going to use this from a new component
